% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Title}
\usage{
birthday(x)
}
\arguments{
\item{x}{Number of people in a random sample}
}
\value{
Probabilit(y/ies) of 2 or more people in sample(s) sharing same birthday
}
\description{
Title
}
\examples{
birthday(20:24)
}
