% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{my first function}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A numeric Vector.}
}
\value{
A list with components 'x' and 'y', where 'y' is the squre of 'x'
}
\description{
my first function
}
\examples{
myfirstfunction(1:10)
}
