% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The average of the Normal Distribution}

\item{sigma}{The standard Deviation of the Normal Deviation}

\item{a}{The value at which the probability calculated from negative infinity goes to}
}
\value{
An image of the distribution, with the region of the probability shaded along with its associated area rounded to decimal points
}
\description{
Title
}
\examples{
myncurve(500,200,a=300)
}
